[rust-loader] Refactor errors into individual modules
